// frontend/src/types/AnnounceType.ts
export type AnnounceType = {
    id: number;
    pet_id: number;
    urgency: boolean;
    org_id: number;
    announce_text: string;
    period: string;
    hidden: boolean;
    admin_id: number | null;
    animaltype: string;
    bloodtype: string;
    organization: string;
    address: string;
    workinghours: string;
    photo: string;
    petname: string;
};